Changed: chaper2 and 3 from old files
