--- src/yunow/chapter3.py
@@ -1,381 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Story: chapter 1 'Depature'
-"""
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-sys.path.append('storybuilder')
-
-from storybuilder.builder import world as wd
-from src.yunow import config as cnf
-THM = cnf.THEMES
-
-# titles
-TITLE = [
-        "mamazonなう",# NOTE: 何でもamazon
-        "魔王退治なう",
-        "全滅なう",# NOTE: 旅立ち編終了、ラストで死に戻り体験
-        ]
-
-# scenes
-## ep11 scene
-def sc_preparation(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam = w.sol, w.mako, w.mother
-    return w.scene("旅立ちの準備だ",
-            h.be(w.stage.myliving, w.day.firstawake2),
-            h.look("$yumarkを無事に落札者に発送し終え",
-                "$Sは空腹を抱えて自宅へと戻ってきた"),
-            sol.talk("それで魔王って奴はどこに住んでるんだよ？"),
-            mam.talk("うちの息子がそんなこと知ってる訳ないじゃないの。", "ねえ$makoはどう思う？"),
-            mako.talk("魔王城だと思いますよ"),
-            h.look("だが中に入って目の前に広がったのは",
-                "赤頭の男$solとピンクのおかっぱ頭の$n_mako",
-                "それに$Sの母親がテーブルの上に広げた色とりどりのお菓子を囲んで",
-                "仲良さそうに談笑する姿だった"),
-            yusha.talk("お前ら何してんの？"),
-            sol.talk("お前も食えよ。", "$makoさんがくれた菓子めちゃめちゃうめえぞ"),
-            mam.talk("あんたお帰り"),
-            mako.talk("あ", "$taroおかえりなさい。", "ちゃんと手続きできましたか？"),
-            h.look("まるで長年連れ添ったパーティのように見えたが",
-                "まだ三人とも昨日顔を合わせただけのただの他人だ。",
-                "それも母親は全く関係ない。", "どちらかといえば$Sの方の知り合いである"),
-            yusha.talk("だからなんで母さんは馴染んでんの！"),
-            mam.talk("お前の友だちなんだから$meの息子や娘みたいなもんじゃない？"),
-            yusha.talk("そういう何でも$meのものみたいな考えはやめてくれ……"),
-            mam.talk("おい", "ちょっと$yusha……"),
-            h.deal("必死に付いてこようとする母親を追い払い",
-                "$Sは$solと$n_makoとお菓子と共に",
-                "自室に入った"),
-            yusha.talk("とりあえず魔王退治をする為には何が必要なのかを",
-                "旅立つ前に考えなきゃいけないと思うんだ"),
-            sol.talk("なんか伝説の武器とか手に入れて",
-                "適当にバーンてやればいいんじゃね？"),
-            h.look("真剣に考える気のない$solを放り出し",
-                "$Sは自分だけをじっと見つめている少女に相談をする"),
-            yusha.talk("なあ$mako。",
-                "まずはちゃんとした仲間が必要だと思うんだ。",
-                "旅は過酷と聞いているし",
-                "何より町の外は強いモンスターがうじゃうじゃしている。",
-                "それに立ち向かうには今の$meではどうしても力不足なんじゃないだろうか"),
-            mako.talk("うんうん。", "$meもそう思います。",
-                "あ", "それよりコレ食べます？"),
-            yusha.talk("ありがとう……うん、うまいな。",
-                "じゃねえ！", "そもそもこの大量のお菓子は何なんだ？"),
-            h.deal("床を叩きつけると", "お菓子を盛った木の大皿が跳ねてそこから宝石のような色とりどりのクッキーが落下した"),
-            mako.talk("ああ、もう。",
-                "もったいないなあ"),
-            h.deal("少し唇を尖らせて彼女は落ちたお菓子を拾う"),
-            mako.talk("$meの朝ご飯。", "折角だから$taroと一緒に食べようと思ってね。",
-                "足りなかったら言ってね。",
-                "いくらでも持ってこさせるから"),
-            yusha.ask("流石にご飯にお菓子はないだろ……てか",
-                    "今何て言った？", "持ってこさせる？"),
-            mako.talk("うん、", "知らない？", "コレ"),
-            h.look("$n_makoは彼女の$phoneを見せた。",
-                "そこには$st_amazonと表示され",
-                "お菓子みたいな食べ物だけでなく",
-                "服や家具", "他に武器や防具なんてものまでが価格と共に映っていた"),
-            yusha.talk("コレ……何？"),
-            mako.talk("$st_amazonよ。",
-                "例えば飲み物が欲しいなあって思うでしょ？",
-                "で、ここからタピオカミルクティー選ぶでしょ？",
-                "当日配送選ぶでしょ？",
-                "はい終わり"),
-            h.deal("$Sも$solも彼女が何をしているのかさっぱり分からないまま「お、おう」と頷く"),
-            h.deal("しかし彼女が中央にピンクの甘いゼリーの乗ったクッキーを一つ食べ終えたところで",
-                "玄関先で声がした"),
-            mam.talk("ねえ$yusha。", "こんなものが届いたんだけど？"),
-            h.look("母親が持ってきたのは小さな紙の箱だった。",
-                "中身はガラスのコップを満たす",
-                "何か小さな粒状の黒いものがころころと入っているのが分かる不思議な薄茶色の液体で",
-                "$n_makoはそれを取り出すとストローを差して美味しそうに飲み始める"),
-            mako.talk("最近これ流行ってるんだってね……美味しい"),
-            h.deal("彼女は笑顔を見せると、"),
-            mako.ask("$taroも飲む？"),
-            h.deal("とそのストローを向けた"),
-            )
-
-def sc_mamazon(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam, daima = w.sol, w.mako, w.mother, w.daimaou
-    return w.scene("mamazonは何でも揃う",
-            yusha.talk("いやそれは要らない"),
-            h.look("即座に断った$Sを驚きの表情で彼女は見つめたが",
-                "もう一度ストローを咥えるとジュルジュルと音をさせながら液体を啜った"),
-            mako.talk("美味しいのに。",
-                "とにかくね$st_amazonなら何でも揃えられるのよ。",
-                "何だったらこれで旅に必要なものも全部揃えてしまえばいいのに"),
-            sol.talk("いくらそのゾマホンてのが凄いって言ったって何でも揃うってこたぁねえだろ？",
-                "それじゃあ……畑を耕すクワは？"),# NOTE:何か簡単なもの
-            h.deal("普通なら注文してから良いものなら納品まで一月かかるものもあるというのに",
-                "すぐにそれは届けられた。", "それも現代の名工と呼ばれる$n_muramasaが仕上げたものだ"),
-            sol.talk("嘘だろ……"),
-            h.look("$solは確かに刻まれた$n_muramasaの紋に言葉を失っていた"),
-            yusha.talk("じゃあ……これならどうだ？"),# NOTE:ちょっと無理め
-            h.deal("$Sは常々欲しいと思っていた等身大$n_idolの抱き枕を注文したが",
-                "オークションにもなかなか出回らないという逸品だ。",
-                "しかしそれもすぐに届いた"),
-            yusha.talk("それじゃあ"),
-            h.deal("$Sと$solは次々と無理難題をふっかけてみたが",
-                "それらは悉く届けられ",
-                "やがて$Sの自室には収まりきらなくなり",
-                "家の前に巨大な精霊像や荷馬車",
-                "最新式のテントまでが揃えられてしまった"),
-            sol.talk("なあ……もうこれで魔王退治すればいいんじゃね？"),
-            h.look("その光景に$solは呆れてそう言ったが、"),
-            mako.talk("魔王退治までは流石に……あ"),
-            h.look("どうやら$n_makoはそれを見つけてしまったらしい"),
-            mako.talk("ねえ。",
-                "これちょっと見てよ……"),
-            h.look("言われて二人は彼女の$phoneを覗き込む。",
-                "そこには『魔王退治』と書かれていた"),
-            yusha.talk("コレって注文したら本当に魔王退治してくれるのか？"),
-            h.look("$Sは嬉々として$n_makoに尋ねたが",
-                "流石の彼女も首を捻った"),
-            sol.talk("まあものは試しだ。", "注文してみようぜ"),
-            h.deal("$solの無駄に前向きな姿勢に背中を押され",
-                "戸惑う彼女に変わって$Sは『注文』を押したのだった"),
-            )
-
-## ep12 scenes
-def sc_defeatmaou(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam, daima = w.sol, w.mako, w.mother, w.daimaou
-    return w.scene("魔王退治",
-            h.be(w.stage.bedroom, w.day.firstawake2),
-            h.look("今", "$Sたちの目の前には『魔王退治』と表示された$phoneがある"),
-            h.explain("$phoneとは不思議な鉱石で出来た板で",
-                "それを持っていれば遠隔地であっても文章や画像のやり取りが可能という魔法の道具なのだが",
-                "更にそれを使って様々な注文ができる$st_amazonというサービスがあり",
-                "$Sとその愉快な仲間二人は",
-                "そこに表示された『注文』に掛けられた彼の指をじっと凝視していた"),
-            yusha.talk("コレって注文したら本当に魔王退治してくれるのか？"),
-            h.think("内心では流石にそれは無理だろう", "と$Sである少年は思っていたのだが",
-                "赤髪の長身男性$solは彼の肩に手を置いて", "こう言った"),
-            sol.talk("まあものは試しだ。", "注文してみようぜ"),
-            mako.talk("ちょっとやめときなさいよ。", "流石に魔王退治を注文するなんてどうかしてる"),
-            h.look("苦言を呈したのはピンクのおかっぱ頭の少女、$n_makoだ。",
-                "元々は彼女が$st_amazonを紹介してくれたのだが",
-                "魔王退治注文はどうも気に入らないらしい"),
-            yusha.talk("これで魔王退治できたら儲けもんだし",
-                "まあ注文するだけ注文してみようぜ"),
-            h.deal("楽観的な$solに賛同した$Sは", "そこに記された金額や注意事項を一切確認することなく",
-                "実に無責任な指の動きで『注文』を押した"),
-            mako.talk("……そもそも魔王退治なんてされたら困るんですけど"),
-            h.look("あーあ", "という溜息を漏らした$n_makoは「何も起こんねえな」と笑う男性二人を横目に",
-                "何故かつまらなさそうだ"),
-            h.be("しばらく待ってみたが何も起こらない"),
-            h.think("いや", "妙な地響きを感じて三人ともそわそわと部屋を見回す"),
-            yusha.talk("な、何だ？"),
-            h.look("その時",
-                "窓の外が強烈に光った"),
-            )
-
-def sc_daimaou(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam, daima = w.sol, w.mako, w.mother, w.daimaou
-    return w.scene("大魔王",
-            mako.talk("何！？"),
-            h.look("そう叫んだのは$n_makoだ。",
-                "今までに見せたことのない神妙な顔つきになって部屋を出ていく"),
-            sol.talk("おい！", "何だってんだよ"),
-            h.move("続いて$Sと$solも外に飛び出て行った"),
-            mam.talk("何なの？", "みんな外が騒がしいわねえ"),
-            h.look("その様子を洗濯物を干しながら母親が見やったが",
-                "今までにない晴天に眩しそうに手を空に翳した"),
-            w.tag.br(),
-            h.look("街中の人間が家の外に出て何やら騒いでいた"),
-            h.look("特に中央広場は街中の人が集まったような混雑ぶりで",
-                "まるでお祭り騒ぎといってもいい賑わいぶりだ"),
-            yusha.talk("何だよありゃ……"),
-            h.look("その騒ぎの中心には城の兵士がいて「押すな！　押すな！」と叫びながら",
-                "何やら札が立てていた"),
-            sol.talk("何だって？"),
-            yusha.talk("どうした$sol"),
-            sol.talk("いや", "なんか『魔王が退治されて平和が戻った』とか書いてあんだが",
-                "どういうことだ？"),
-            mako.talk("え？"),
-            h.hear("その声をあげたのは$n_makoだった"),
-            mako.talk("退治とか、そんなの有り得ないのに……"),
-            yusha.talk("まあホントに退治されたんなら", "ありがたいじゃないか。",
-                "魔王退治の旅に出なくても済むし"),
-            h.deal("$Sは喜ぶ町の人々の顔を眺めやり",
-                "まるで魔王退治を自分が成し遂げたことのような満足感に浸っていた"),
-            h.look("だがその隣で何故か$n_makoは悲しそうな表情を浮かべている"),
-            yusha.talk("何だよ$mako。",
-                "そんなに$meと旅したかったならこれから平和になった世界を旅すればいいじゃないか"),
-            h.deal("そう声を掛けるが", "表情は変わらない"),
-            h.look("と、空が急に暗くなる"),
-            h.look("黒雲に覆われ", "一気に夜に変わったかと思うほどの闇に襲われ",
-                "かと思えばそこがメリメリと裂けるようになって",
-                "真っ赤に空が燃え始めた。",
-                "その天の炎が柱となり、"),
-            yusha.talk("あ……"),
-            h.look("城へと突き刺さった"),
-            yusha.talk("何だありゃ……"),
-            h.look("城は一瞬で半壊し", "燃え上がる"),
-            h.think("その場の誰もが何が起こったのか分からず",
-                "ただ叫び声や驚きの声を上げていたが",
-                "それらの喧騒を一気に消し飛ばすような地響きにも似た声が響き渡った"),
-            daima.talk("魔王は先ほど倒された。",
-                "その代わりに今後は$meが世界を恐怖に震え上がらせてやろう。",
-                "我が名は大魔王。",
-                "世界に絶望と混沌をもたらす最凶の闇"),
-            h.explain("その日", "世界は闇に覆われた"),
-            yusha.talk("まじかよ……$st_amazonすげえな"),
-            sol.talk("感心するとこソコじゃねえだろ！"),
-            )
-
-## ep13 scenes
-def sc_destruction(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam, daima = w.sol, w.mako, w.mother, w.daimaou
-    return w.scene("全滅",
-            h.be(w.stage.hometown, w.day.firstawake2),
-            yusha.talk("大魔王……"),
-            h.look("誰もが空を見上げ",
-                "どこから響いてくるとも知れない声に震え上がっていた"),
-            sol.talk("何だよ大魔王って……でっかい魔王になったのか？"),
-            h.look("事態を理解できていない赤髪の長身男性$solを一瞥して",
-                "$Sである少年は今一度", "天を見上げる。",
-                "彼の隣でピンクのおかっぱ頭の$n_makoは",
-                "不安からか$Sのシャツの袖をぎゅっと握って体を寄せている"),
-            h.hear().t("大魔王って何だよ"),
-            h.hear().t("せっかく魔王退治されたと思ったら", "これじゃ無意味じゃねえかよ"),
-            h.hear().t("このままじゃまた怯える毎日に逆戻りじゃないの！"),
-            h.deal("中央広場に集まった人々は最初こそ声を失っていたが徐々に騒ぎ出し、"),
-            h.hear().t("死にたくねえ！"),
-            h.deal("誰かが声を上げたのを合図に散り散りになって広場から逃げ出していく。",
-                "みなが自分の家へと閉じ籠もるようで",
-                "次々とドアが閉められ", "窓も閉じてカーテンまで引かれてしまった"),
-            sol.talk("魔王ですら無理なのに大魔王って何だよ！"),
-            h.hear("$solが叫ぶ"),
-            mako.talk("……こんなの", "絶対おかしい"),
-            h.look("$n_makoはぶつぶつと言いながら何か考え込んでいるらしく",
-                "$Sの袖を離すと自分の$phoneを出して何かし始めた"),
-            h.explain("$phoneとは同じ道具を持つ者ならどんな遠隔地であっても文章や画像をやり取りできる",
-                "魔法の道具のことだ"),
-            h.look("$Sも自分の$phoneで他の国の人々の様子を見てみたが",
-                "誰もが大魔王の存在に驚き", "外出ばかりか", "何か言葉を発することすら自粛しているようだった"),
-            yusha.talk("今度は大魔王かよ……こんなの無理だよな"),
-            h.look("そう呟いた$Sは",
-                "何故か笑みを浮かべていて",
-                "寧ろ大魔王の出現を楽しんでいるかのような得意げな表情になっている"),
-            h.think("だがその彼が得意げに胸を張っていられたのもごく僅かな間だけだった"),
-            sol.talk("おい……ありゃ何だ？"),
-            h.look("$solが指したのは遥か遠くの山の上だ。",
-                "そこに小さな黒い靄のようなものが生まれ",
-                "どうやらそれは徐々に大きくなっている"),
-            mako.talk("ちょっと待ってね。",
-                "今確認してみるから"),
-            h.look("$n_makoはそう言って$phoneを向ける"),
-            mako.talk("拡大したら何か分かるかも"),
-            h.look("切り取ったその風景が$phoneに写し取られ", "彼女の指の動きでどんどん巨大になっていく。",
-                "まるで虫眼鏡を何重にも重ねたみたいだと$Sは思ったが",
-                "実際にそれをするとぼやけて見えなくなるだけだった"),
-            mako.talk("これって……"),
-            sol.talk("ああ……"),
-            h.look("それはどう見ても大量の翼を持ったモンスターたちだった。",
-                "鳥型のものもいるが", "それらの背後には$Sが$gameの中でだけ遭遇したあの伝説のドラゴンらしき影も映っていた"),
-            sol.talk("おい。",
-                "こんなの来たらただじゃ済まねえぞ$yusha！"),
-            yusha.talk("……既にただじゃ済んでないみたいだな"),
-            h.look("そう言って$Sは二人に自分の$phoneを見せる。",
-                "そこに表示されていたのは――"),
-            h.look().emphasis("全滅なう"),
-            h.look("次々と阿鼻叫喚の断末魔の如くに『全滅なう』が流れていく"),
-            mako.talk("次々に町が壊滅してるのね……なんてことを"),
-            )
-
-def sc_theend(w: wd.World):
-    h = yusha = w.yusha
-    sol, mako, mam, daima = w.sol, w.mako, w.mother, w.daimaou
-    return w.scene("そして終わり？",
-            h.look("その断末魔はやがて近隣の町や村にまで迫ってくる"),
-            sol.talk("どうすんだよ", "なあ$yusha"),
-            yusha.talk("どうするって……やるしかないんじゃね？",
-                "だって$me", "$Sだし"),
-            h.look("そう言って親指を立てた$Sに$solはニヤリとして「そうだな」と答える"),
-            mako.talk("あんたらバカじゃないの？",
-                "仮に全員魔王だったとして", "あれだけの数相手にどうにかできるなんてのは妄想でしかないわよ"),
-            h.look("$n_makoは辛辣な意見と共に鋭い目線を二人に投げる。",
-                "既に広場から人はいなくなり",
-                "彼らだけが立っていた"),
-            h.look("そこに", "奇妙なざわざわとした音が迫ってくる"),
-            yusha.talk("え……この音って"),
-            mako.talk("そうよ。", "やつらの羽音よ"),
-            h.look("まだ遠くの空が黒い靄で覆われているだけなのに",
-                "それはどんどん大きくなってくる"),
-            sol.talk("$meちっさい頃から虫だけはダメなんだわ。",
-                "悪い$yusha。", "逃げるわ"),
-            yusha.talk("は？", "何言ってんだよ$sol！", "おいちょっと待てって！"),
-            h.deal("だが伸ばした手は届かず",
-                "赤髪の大男はあっという間に駆けていってしまった"),
-            yusha.talk("何だよあいつは……晩メシおごってやった恩を忘れたのかよ！"),
-            h.be("$n_makoを見たが", "どうやら彼女は$solのように逃げるつもりはないらしい"),
-            yusha.talk("$makoもいいんだぞ"),
-            mako.talk("$meは常に$taroと共にありますから……それに",
-                "ちょっと確かめたいことがあって"),
-            h.deal("そう答えた$n_makoは自分の$phoneで必死に何かしている"),
-            h.look("いよいよ羽音が空気を満たすほどとなり",
-                "燃えている空が黒い絨毯によって覆われた"),
-            yusha.talk("来るなら来い！"),
-            h.think("何故だろう。",
-                "この時の彼は恐れや怯えといった感情とは無関係だった。",
-                "ただ剣を構えて自分がこの世界を救うというその一心だけが支配していた"),
-            h.hear("そんな二人に向け",
-                "どのモンスターのものかは分からない唸り声が響き",
-                "一気に黒い空が落ちてくる"),
-            h.deal("やってやるぜと意気込んだ$Sの気持ち虚しく",
-                "圧倒的な数の暴力によって彼は剣を振るうことすら出来ずにそのまま意識ごと呑み込まれてしまった"),
-            h.deal("全滅なう"),
-            h.deal("誰かがそれを見て$phoneで呟き", "ケタケタと笑い声をあげたのだった"),
-            )
-
-# episodes
-def ep11(w: wd.World):
-    return (w.chaptertitle(TITLE[0]),
-            sc_preparation(w),
-            sc_mamazon(w),
-            )
-
-def ep12(w: wd.World):
-    return (w.chaptertitle(TITLE[1]),
-            sc_defeatmaou(w),
-            sc_daimaou(w),
-            )
-
-def ep13(w: wd.World):
-    return (w.chaptertitle(TITLE[2]),
-            sc_destruction(w),
-            sc_theend(w),
-            )
-
-# outline
-def story_baseinfo(w: wd.World):
-    return [
-            ]
-
-def story_outline(w: wd.World):
-    return [
-            ]
-
-# main
-def story(w: wd.World):
-    return (w.maintitle(cnf.TITLE["chap1"]),
-            ep11(w),
-            ep12(w),
-            ep13(w),
-            )
-
-def main(): # pragma: no cover
-    from src.yunow.story import world
-    w = world()
-    return w.build(story(w))
-
-
-if __name__ == '__main__':
-    import sys
-    sys.exit(main())
-
