+++ src/yunow/config.py
@@ -139,5 +139,7 @@ MOTIFS = {
 TITLE = {
         "chap1": "第一章　旅立ち編",
         "chap2": "第二章　冒険初心者編",
-        "chap3": "第三章　さよなら故郷編",
+        "chap3": "第三章　ゴブリン討伐編",
+        "chap4": "第四章　さよなら故郷編",
+        "chap5": "第五章　新大陸ローレア編",
         }
