+++ src/yunow/story.py
@@ -9,8 +9,8 @@ sys.path.append('storybuilder')
 from storybuilder.builder import world as wd
 from src.yunow import config as cnf
 from src.yunow import chapter01 as chap1
-from src.yunow import chapter4 as chap4
-from src.yunow import chapter5 as chap5
+from src.yunow import chapter02 as chap2
+from src.yunow import chapter03 as chap3
 THM = cnf.THEMES
 
 
@@ -22,8 +22,8 @@ def story_baseinfo(w: wd.World):
     return [
             ("story", story(w), w.yusha, w.yusha),
             ] + chap1.story_baseinfo(w) \
-                + chap4.story_baseinfo(w) \
-                + chap5.story_baseinfo(w)
+                + chap2.story_baseinfo(w) \
+                + chap3.story_baseinfo(w)
 
 def story_outline(w: wd.World):
     return [
@@ -34,8 +34,8 @@ def story_outline(w: wd.World):
                 w.yusha.go(w.i.trouble),
                 True),
             ] + chap1.story_outline(w) \
-                + chap4.story_outline(w) \
-                + chap5.story_outline(w)
+                + chap2.story_outline(w) \
+                + chap3.story_outline(w)
 
 # main
 def world():
@@ -52,8 +52,8 @@ def world():
 def story(w: wd.World):
     return (w.maintitle("勇者なう！"),
             chap1.story(w),
-            chap4.story(w),
-            chap5.story(w),
+            chap2.story(w),
+            chap3.story(w),
             )
 
 def main(): # pragma: no cover
